Update README
With a variety of new additions for this release, such as Tippy, KaTeX, and multiple new chart implementations, a major revision for the README was necessary.
Further optimizations can likely be made to reduce the overall length of the page, but this is a good place to start.
Also added some additional tooling in our tech stack that was previously unmentioned.